# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recall API",
    "description": "Recall \u2014 AI-powered second brain. Ingest links, voice notes, PDFs and images via Telegram. Search, map, and quiz your knowledge via the web dashboard.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/telegram": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Telegram login verification",
        "description": "Verifies the signature of the Telegram login widget data and sets the recall_session JWT cookie.",
        "operationId": "auth_telegram_auth_telegram_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid signature or expired authentication data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/google/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Google OAuth callback",
        "description": "Processes the OAuth callback, exchanges authorization code, and stores encrypted refresh token.",
        "operationId": "auth_google_callback_auth_google_callback_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "OAuth code exchange failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Programmatic error code."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error description."
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "LoginStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Authentication status."
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "LoginStatusResponse"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "recall_session",
        "description": "JWT stored in the httpOnly 'recall_session' cookie."
      },
      "telegramInitData": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Telegram Web App initData in Authorization header (format: TelegramInitData <init_data>)."
      }
    }
  }
}